Move types and store to separate files
